Add Transition Between Chords and Fast Riff
The speed change between the part after the chords and the fast
parts is a bit jarring. Introduce something intermediate: something
that starts with smaller chords a picks up speed. Then rock out.

Note, the transition is interesting but not finished. There doesn't
seem to be a clean transition to the between my mid-tempo
part as well. I do like the unexpected dissonance that the transition
(which uses multiple chord types) introduces. But how do I get
to the speed? Perhaps a second transition would help, with a
rhythm pattern that emulates the upcoming shredding.